memory::memory (closes #21) and memory::swap (closes #23) for Windows. Underline type for memory::Information was changed from the usize to u64 on all targets

+++ heim-memory/src/lib.rs
@@ -19,5 +19,13 @@ pub use self::swap::*;
 /// Re-exported measurement units used in this crate.
 pub mod units {
     pub use heim_common::units::iec::information::*;
-    pub use heim_common::units::iec::usize::Information;
+    // The obvious choice is too use `usize` type as a memory size representation.
+    // Yet, the problem is `Information` base unit is "bit", not a "byte",
+    // therefore all the time we are having the multiplication by 8,
+    // and this breaks `x86` systems, because even small x86 VMs
+    // have so much memory bits, that it overflows their `usize`
+    // (which is `u32` in that case).
+    //
+    // Therefore, we are stuck with a `u64` for all systems.
+    pub use heim_common::units::iec::u64::Information;
 }
